new tap values
we changed the tap values to 24, 32, 45, 60, 125, 200 while changing the mux frequency to 800 and the time stamp to output correctly (It was pulling the period variable from somewhere besides sensors.c).  We also included the stutter code as well as changed the impedance equation to match our new method with using adc value instead of gain.  This is set up for the callibration of board #4 currently.