Brief Explanation of Code to improve Documentation
I added comments to almost every line of code to explain its functionality as well as added additional comments in the beginning to explain I2C communication and the purpose of this code.  I also got rid of unnecessary lines of code.  I also switched the name of the function replace_this_mess to muxPinReset in order to improve readability.