NEW : 250 tap values, 4 adc reads per dactimercallback, MUX on/off adc read, new ImpedanceCalc file, and AutoCal code implemented (CODE IN AUTOCAL SETUP)
First major change to this code is the implementation of 250 tap values instead of the 6 we were using previously.  The old 6 tap values are still commented in but the 250 are being used.  We had to adjust the matlab code used for AUTOMATION CALIBRATION to adapt to this change.  (this is still untested but can't be done on Jared's MBP). We the adjusted the MUX to turn on and off in order to control the signal instead of directly turning on and off the voltage/current going into the sensors every time we wanted an adc read.  This allowed easy changes to the number of adc reads per dactimercallback. Lastly we added the impedanceCalc file in order to clean up the sensors.c.  In our callibration process for 250 tap values we will generate around 500 lines of code used for calibration that will need to be copied in for every distinct board.  Having a separate file for this makes it an easier process.  We will also get our high and low bounds for each tap value from this calibration process as well.