Changed ImpedanceCalc to switch statement. Simplified code to conserve memory, updated 250 tap values and adc read to perform better
I took out a few arrays that were clogging up the RAM as well as simplified some for loops that were unnecessary.  I also rearranged the adc read to read from every sensor four times before outputting the average of all four reads for a specific sensor.  It all works.  Still working on how to get more accurate reads by adjusting the high and low cuts ranges for tap values as well as creating a PI controller to account for the time it takes to adjust the tap value.