48 Hour code and V1.31 code additions
48 hour code has been added.  It is at a point where the user needs to start and stop recording using a mobile device every time the PCB is turned off and on.  This was done by adding OSAL_SNV = 2 to the built in variables, enabling flash storage.  Other additions were made in sensors.c.  A boolian titled FOURTYEIGHT turns on the fourty eight code. This needs to be fixed so that the fourty eight hour code runs seemlessly and stores the current_sector variable while the PCB turns on and off. These modifications were made in simple_peripheral.c lines 700 and on.

Changes were also made for V1.31 of the PCB.  VONETHREE is the boollian variable that sets the code up for v1.31.  This is a work in progress and still needs a lot of help.  By setting VONETHREE to true, the code sets the I2C registers for the new digipot and DAC that were replaced in V1.31.  It also sets the DAC around lines 360. (this is where I would look first to improve this.  Current output shows signs of inaccurate communication with the DAC.