48 Hour Finished /EMG in progress
Updated the 48 hour code to begin writing to the SD card only in factors of 33.

Sensors.c is cleaned up. EMG code is updated so that if EMG is true the code will collect an adcread every case in the finite state machine.  Still working on doubling the frequency.  Currently the EMG code is sampling at 150Hz per sensor and the impedance code samples at 50hz per sensor.