Updated to run EMG and Impedance
EMGIMP set to true runs EMG and Impedance code.  NUM_CYCLES PER EMG OUTPUT is the variable that should be compared with NUM CYCLES PER OUTPUT (impedance output) when setting the frequencies of each (i.e. how often we are sampling EMG and impedance).  The standard that I know for sure works is what is in this code.