EMGIMP working. 48 hour code starts at sector 10000
I created a function sensors_serializer_output and added it to every case of the fsm.  This allowed for the code to output EMG values for every sensor instead of skipping a lot of sensors.  This function conntains all the code necessary to add to the serializer and output to the uart/sd card if needed.